Namespace and folder refactor
