+++ samples/AuthenticationOptionsSample/Views/_ViewImports.cshtml
@@ -1,2 +1,2 @@
-﻿@using Finbuckle.MultiTenant.Core
+﻿@using Finbuckle.MultiTenant
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
