+++ src/Finbuckle.MultiTenant.AspNetCore/Extensions/ServiceCollectionExtensions.cs
@@ -13,13 +13,14 @@
 //    limitations under the License.
 
 using System;
+using Finbuckle.MultiTenant;
 using Finbuckle.MultiTenant.AspNetCore;
 using Finbuckle.MultiTenant.Core;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
-namespace Microsoft.Extensions.DependencyInjection
+namespace Finbuckle.MultiTenant
 {
     public static class ServiceCollectionExtensions
     {
