+++ src/Finbuckle.MultiTenant.AspNetCore/Internal/Constants.cs
@@ -13,6 +13,7 @@
 //    limitations under the License.
 
 using System;
+
 namespace Finbuckle.MultiTenant.AspNetCore
 {
     /// <summary>
