+++ src/Finbuckle.MultiTenant.AspNetCore/Internal/InMemoryMultiTenantStoreOptions.cs
