+++ src/Finbuckle.MultiTenant.AspNetCore/Internal/MultiTenantOptionsFactory.cs
