+++ src/Finbuckle.MultiTenant.AspNetCore/Internal/MultiTenantOptionsManager.cs
