+++ src/Finbuckle.MultiTenant.AspNetCore/MultiTenantBuilder.cs
@@ -13,22 +13,18 @@
 //    limitations under the License.
 
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
-using Finbuckle.MultiTenant.Core;
-using Finbuckle.MultiTenant.Core.Abstractions;
-using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Logging;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.OpenIdConnect;
-using Microsoft.AspNetCore.Authentication.OAuth;
+using Finbuckle.MultiTenant.AspNetCore;
+using Finbuckle.MultiTenant.Strategies;
+using Finbuckle.MultiTenant.Stores;
 
-namespace Finbuckle.MultiTenant.AspNetCore
+namespace Finbuckle.MultiTenant
 {
     /// <summary>
     /// Configures <c>Finbuckle.MultiTenant.AspNetCore</c> services and configuration.
