+++ src/Finbuckle.MultiTenant.AspNetCore/Strategies/BasePathMultiTenantStrategy.cs
@@ -13,14 +13,11 @@
 //    limitations under the License.
 
 using System;
-using System.Threading.Tasks;
-using Finbuckle.MultiTenant.AspNetCore;
 using Finbuckle.MultiTenant.Core;
-using Finbuckle.MultiTenant.Core.Abstractions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
-namespace Finbuckle.MultiTenant.AspNetCore
+namespace Finbuckle.MultiTenant.Strategies
 {
     public class BasePathMultiTenantStrategy : IMultiTenantStrategy
     {
