+++ src/Finbuckle.MultiTenant.AspNetCore/Strategies/RemoteAuthenticationMultiTenantStrategy.cs
@@ -14,9 +14,6 @@
 
 using System;
 using System.Linq;
-using System.Threading.Tasks;
-using Finbuckle.MultiTenant.Core;
-using Finbuckle.MultiTenant.Core.Abstractions;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.OAuth;
 using Microsoft.AspNetCore.Authentication.OpenIdConnect;
@@ -25,7 +22,7 @@ using Microsoft.AspNetCore.Http.Features;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 
-namespace Finbuckle.MultiTenant.AspNetCore
+namespace Finbuckle.MultiTenant.Strategies
 {
     public interface IRemoteAuthenticationMultiTenantStrategy
     {
