+++ src/Finbuckle.MultiTenant.Core/Internal/Utilities.cs
