+++ src/Finbuckle.MultiTenant.Core/MultiTenantException.cs
@@ -14,7 +14,7 @@
 
 using System;
 
-namespace Finbuckle.MultiTenant.Core
+namespace Finbuckle.MultiTenant
 {
     /// <summary>
     /// A derived <c>Exception</c> class for any exception generated by <c>Finbuckle.MultiTenant</c>.
