+++ src/Finbuckle.MultiTenant.EntityFrameworkCore/Internal/MultiTenantIdentitySupportClasses.cs
