+++ test/Finbuckle.MultiTenant.AspNetCore.Test/BasePathMultiTenantStategyShould.cs
@@ -13,8 +13,10 @@
 //    limitations under the License.
 
 using System;
+using Finbuckle.MultiTenant;
 using Finbuckle.MultiTenant.AspNetCore;
 using Finbuckle.MultiTenant.Core;
+using Finbuckle.MultiTenant.Strategies;
 using Microsoft.AspNetCore.Http;
 using Moq;
 using Xunit;
