+++ test/Finbuckle.MultiTenant.AspNetCore.Test/RouteMultiTenantStrategyShould.cs
@@ -17,6 +17,7 @@ using System.Collections.Concurrent;
 using Finbuckle.MultiTenant;
 using Finbuckle.MultiTenant.AspNetCore;
 using Finbuckle.MultiTenant.Core;
+using Finbuckle.MultiTenant.Strategies;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Moq;
