+++ test/Finbuckle.MultiTenant.Core.Test/InMemoryMultiTenantStoreShould.cs
@@ -13,8 +13,8 @@
 //    limitations under the License.
 
 using System;
-using System.Collections.Concurrent;
-using Finbuckle.MultiTenant.Core;
+using Finbuckle.MultiTenant;
+using Finbuckle.MultiTenant.Stores;
 using Xunit;
 
 public class InMemoryMultiTenantStoreShould
