+++ test/Finbuckle.MultiTenant.Core.Test/StaticMultiTenantStrategyShould.cs
@@ -14,7 +14,9 @@
 
 using System;
 using System.Collections.Concurrent;
+using Finbuckle.MultiTenant;
 using Finbuckle.MultiTenant.Core;
+using Finbuckle.MultiTenant.Strategies;
 using Xunit;
 
 public class StaticTenantResolverShould
