+++ test/Finbuckle.MultiTenant.EntityFrameworkCore.Test/TestEntitities.cs
@@ -17,6 +17,7 @@ using Microsoft.EntityFrameworkCore;
 using Finbuckle.MultiTenant.EntityFrameworkCore;
 using Finbuckle.MultiTenant.Core;
 using System.ComponentModel.DataAnnotations;
+using Finbuckle.MultiTenant;
 
 public class TestDbContext : MultiTenantDbContext
 {
